+++ src/Symfony/Component/Messenger/Bridge/MongoDb/CHANGELOG.md
@@ -5,3 +5,5 @@ CHANGELOG
 ---
 
  * Introduce the MongoDB Messenger bridge
+ * Add Change Streams support to reduce the latency of idle workers, with the
+   `wait_time` transport option (0 or less disables it)